+++ packages/apollo-server-env/src/polyfills/Object.entries.ts
@@ -1,0 +1,5 @@
+if (!global.Object.entries) {
+  global.Object.entries = function(object: any) {
+    return Object.keys(object).map(key => [key, object[key]] as [string, any]);
+  };
+}