+++ packages/apollo-server-lambda/src/ApolloServer.ts
@@ -44,7 +44,7 @@ export class ApolloServer extends ApolloServerBase {
   }
 
   public createHandler({ cors }: CreateHandlerOptions = { cors: undefined }) {
-    const corsHeaders = {};
+    const corsHeaders: lambda.APIGatewayProxyResult['headers'] = {};
 
     if (cors) {
       if (cors.methods) {
@@ -119,8 +119,13 @@ export class ApolloServer extends ApolloServerBase {
       if (this.playgroundOptions && event.httpMethod === 'GET') {
         const acceptHeader = event.headers['Accept'] || event.headers['accept'];
         if (acceptHeader && acceptHeader.includes('text/html')) {
+          const path =
+            event.path ||
+            (event.requestContext && event.requestContext.path) ||
+            '/';
+
           const playgroundRenderPageOptions: PlaygroundRenderPageOptions = {
-            endpoint: event.requestContext.path,
+            endpoint: path,
             ...this.playgroundOptions,
           };