+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All of the packages in the `apollo-server` repo are released with the same versi
 - express, koa: remove next after playground [#1436](https://github.com/apollographql/apollo-server/pull/1436)
 - Hapi: Pass the response toolkit to the context function. [#1407](https://github.com/apollographql/apollo-server/pull/1407)
 - update apollo-engine-reporting-protobuf to non-beta [#1429](https://github.com/apollographql/apollo-server/pull/1429)
+- Lambda: Look in event.path first when picking endpoint for GraphQL Playground [#1527](https://github.com/apollographql/apollo-server/pull/1527)
 
 ### rc.10
 
@@ -25,7 +26,7 @@ Unused RC, due to publication issues
 - export `bodyParser.Options` to koa [#1334](https://github.com/apollographql/apollo-server/pull/1334)
 - add and use playground in ApolloServer constructor [#1297](https://github.com/apollographql/apollo-server/pull/1297)
 - **breaking**: remove calculate headers as function [#1337](https://github.com/apollographql/apollo-server/pull/1337)
-- **breaking**: remove `fromatParams` [#1331](https://github.com/apollographql/apollo-server/pull/1331)
+- **breaking**: remove `formatParams` [#1331](https://github.com/apollographql/apollo-server/pull/1331)
 
 ### rc.7